+++ client/data/diary.json
@@ -183,8 +183,7 @@
       "tags": [
         "df"
       ],
-      "id": 9,
-      "like": false
+      "id": 9
     },
     {
       "nickname": "dlwlrma",
@@ -432,274 +431,6 @@
       "year": "2020",
       "month": "09",
       "day": "30"
-    },
-    {
-      "nickname": "dlwlrma",
-      "title": "dfdsdfsdf",
-      "year": "2022",
-      "month": "12",
-      "day": "30",
-      "weather": "안개",
-      "imageList": [
-        {}
-      ],
-      "question": "이번 여행에서 맛있게 먹은 음식은 무엇인가요?",
-      "counter": 0,
-      "diary": "ddsfsdfsdfsdfsdf",
-      "price": "15000",
-      "selected": "울산광역시",
-      "city": "북구",
-      "tags": [
-        "테스트"
-      ],
-      "id": 23
-    },
-    {
-      "title": "sdf",
-      "year": 2021,
-      "month": "10",
-      "day": "29",
-      "weather": "구름적음",
-      "diary": "sdfs",
-      "price": 1500000,
-      "selected": "충청북도",
-      "nickname": "dlwlrma",
-      "question": "여행에서 자주 들은 노래는 무엇인가요?",
-      "counter": 1,
-      "tags": [
-        "dfs",
-        "ges"
-      ],
-      "id": 24
-    },
-    {
-      "title": "sdf",
-      "year": 2021,
-      "month": "10",
-      "day": "29",
-      "weather": "구름적음",
-      "diary": "sdfs",
-      "price": 1500000,
-      "selected": "충청북도",
-      "nickname": "dlwlrma",
-      "question": "여행에서 자주 들은 노래는 무엇인가요?",
-      "counter": 1,
-      "tags": [
-        "dfs",
-        "ges"
-      ],
-      "id": 25
-    },
-    {
-      "title": "sdfsdf",
-      "year": 2022,
-      "month": 10,
-      "day": 29,
-      "weather": "황사",
-      "diary": "sdgsgs",
-      "price": 5555,
-      "selected": "경상남도",
-      "nickname": "dlwlrma",
-      "question": "이번 여행 중 나만 알고 싶은 사진 명소는 어디인가요?",
-      "counter": 1,
-      "tags": [
-        "sdgs",
-        "dgsdgs"
-      ],
-      "id": 26
-    },
-    {
-      "title": "sdfsdf",
-      "year": 2022,
-      "month": 10,
-      "day": 29,
-      "weather": "맑음",
-      "diary": "sdgsgs",
-      "price": 5555,
-      "selected": "경상남도",
-      "nickname": "dlwlrma",
-      "question": "이번 여행 중 나만 알고 싶은 사진 명소는 어디인가요?",
-      "counter": 1,
-      "tags": [
-        "sdgs",
-        "dgsdgs"
-      ],
-      "id": 27
-    },
-    {
-      "title": "sds",
-      "year": 2017,
-      "month": 8,
-      "day": 23,
-      "weather": "천둥번개",
-      "diary": "sdbsdbsdb",
-      "price": 5555,
-      "selected": "서울특별시",
-      "nickname": "dlwlrma",
-      "question": "이번 여행에서 가장 귀찮았던 일은 무엇인가요?",
-      "counter": 1,
-      "tags": [
-        "dsfsg"
-      ],
-      "id": 28
-    },
-    {
-      "title": "fdf",
-      "year": 2020,
-      "month": 11,
-      "day": 29,
-      "weather": "안개",
-      "diary": "dfd",
-      "price": 100000000,
-      "selected": "충청북도",
-      "nickname": "dlwlrma",
-      "question": "이번 여행의 주제곡을 골라보자면 무엇일까요?",
-      "counter": 1,
-      "tags": [
-        "sfd"
-      ],
-      "id": 29
-    },
-    {
-      "title": "fdf",
-      "year": 2020,
-      "month": 11,
-      "day": 29,
-      "weather": "안개",
-      "diary": "dfd",
-      "price": 100000000,
-      "selected": "충청북도",
-      "nickname": "dlwlrma",
-      "question": "이번 여행의 주제곡을 골라보자면 무엇일까요?",
-      "counter": 1,
-      "tags": [
-        "sfd"
-      ],
-      "id": 30
-    },
-    {
-      "title": "fdf",
-      "year": 2020,
-      "month": 11,
-      "day": 29,
-      "weather": "안개",
-      "diary": "dfd",
-      "price": 100000000,
-      "selected": "충청북도",
-      "nickname": "dlwlrma",
-      "question": "이번 여행의 주제곡을 골라보자면 무엇일까요?",
-      "counter": 1,
-      "tags": [
-        "sfd"
-      ],
-      "id": 31
-    },
-    {
-      "title": "fdf",
-      "year": 2020,
-      "month": 11,
-      "day": 29,
-      "weather": "안개",
-      "diary": "dfd",
-      "price": 100000000,
-      "selected": "충청남도",
-      "nickname": "dlwlrma",
-      "question": "이번 여행의 주제곡을 골라보자면 무엇일까요?",
-      "counter": 1,
-      "tags": [
-        "sfd"
-      ],
-      "id": 32
-    },
-    {
-      "title": "wjnrtq",
-      "year": 2018,
-      "month": 9,
-      "day": 28,
-      "weather": "황사",
-      "diary": "dfgdfhshdfhshshfdhs",
-      "price": 7777777,
-      "selected": "세종특별자치시",
-      "nickname": "dlwlrma",
-      "question": "다음에 이곳에 다시 여행을 온다면 누구와 함께 오고 싶나요?",
-      "counter": 2,
-      "tags": [
-        "xfdf"
-      ],
-      "id": 33
-    },
-    {
-      "title": "fhwj",
-      "year": 2021,
-      "month": 10,
-      "day": 27,
-      "weather": "맑음",
-      "diary": "fjej",
-      "price": 73737,
-      "selected": "강원도",
-      "nickname": "dlwlrma",
-      "question": "최고의 여행 명소는 어디였나요?",
-      "counter": 1,
-      "tags": [
-        "dfgdn"
-      ],
-      "id": 34
-    },
-    {
-      "title": "dfbwdfhsh",
-      "year": 2020,
-      "month": 8,
-      "day": 27,
-      "weather": "흐림",
-      "diary": "dfhsfhsfhsdfh",
-      "price": 922222222,
-      "selected": "전라북도",
-      "city": "고창군",
-      "nickname": "dlwlrma",
-      "question": "이번 여행에서 가장 특별했던 만남은 무엇이었나요?",
-      "counter": 1,
-      "tags": [
-        "fsdfs",
-        "eghwh"
-      ],
-      "id": 35
-    },
-    {
-      "title": "테스트 중입니다",
-      "year": 2016,
-      "month": 9,
-      "day": 26,
-      "weather": "흐림",
-      "diary": "dfherhdfhsd",
-      "price": 500000,
-      "selected": "충청남도",
-      "city": "청양군",
-      "nickname": "dlwlrma",
-      "question": "최고의 여행 명소는 어디였나요?",
-      "counter": 1,
-      "tags": [
-        "sdfsdf"
-      ],
-      "id": 36
-    },
-    {
-      "title": "마지막 유효성 검사",
-      "year": 2020,
-      "month": 10,
-      "day": 20,
-      "weather": "천둥번개",
-      "diary": "마지막 유효성 검사마지막 유효성 검사마지막 유효성 검사마지막 유효성 검사마지막 유효성 검사마지막 유효성 검사마지막 유효성 검사마지막 유효성 검사마지막 유효성 검사ㅍ",
-      "price": 5000000,
-      "selected": "경상북도",
-      "city": "고령군",
-      "nickname": "dlwlrma",
-      "question": "최고의 여행 명소는 어디였나요?",
-      "counter": 1,
-      "tags": [
-        "ㄴㅇㄹㄴ",
-        "ㅈㅎㅈㅎㅈ"
-      ],
-      "id": 37
     }
   ]
 }
