+++ client/src/pages/MainPage.js
@@ -5,7 +5,8 @@ import { MainTab } from "../components/Main/MainTab";
 import { useSelector } from "react-redux";
 import { SET_TOKEN, DELETE_TOKEN } from "../redux/store/Auth";
 import LoginHeader from "../components/Header/LoginHeader";
-import { useState } from "react";
+import { useState, useEffect } from "react";
+import axios from "axios";
 
 const Main = styled.div`
   display: flex;
@@ -21,12 +22,44 @@ export default function MainPage() {
   console.log("accesstoken", accesstoken);
 
   const [selected, setSelected] = useState("등록순");
+  const [diaryList, setDiaryList] = useState([]);
+  const [loading, setLoading] = useState(false);
+
+  useEffect(() => {
+    // axios.get(`http://localhost:4000/diary/${id}`).then((res) => {
+    setLoading(true);
+    axios.get("http://localhost:4000/diary").then((res) => {
+      const timer = setTimeout(() => {
+        console.log(res.data);
+        let response = res.data;
+        setDiaryList(response.slice(0, 12)); // 받아온 데이터에서 12개만 먼저 result state에 저장
+        response = response.slice(12);
+        setResult(response); // 저장한 데이터 모두 저장
+        setLoading(false);
+      }, 2000);
+      return () => clearTimeout(timer);
+    });
+  }, []);
+
+  // yerin
+  const [hasMore, setHasMore] = useState(true);
+  const [result, setResult] = useState([]);
   return (
     <Main>
       {user.payload.userReducer.isLogin ? <LoginHeader /> : <Header />}
       <div>
-        <MainTab selected={selected} setSelected={setSelected} />
-        <Card selected={selected} />
+        <MainTab selected={selected} setSelected={setSelected} diaryList={diaryList} />
+        <Card
+          selected={selected}
+          diaryList={diaryList}
+          setDiaryList={setDiaryList}
+          hasMore={hasMore}
+          setHasMore={setHasMore}
+          result={result}
+          setResult={setResult}
+          loading={loading}
+          setLoading={setLoading}
+        />
       </div>
     </Main>
   );
